Sockets are now created through NetAddr
Because NetAddr need more control of creation of the socket

Unix sockets can be refered as unix:file:permissions, without ://. The
new form also doesn't require absolute path. The path is converted to
absoluted during creation of NetAddr

+++ src/simpleServer/threadPoolAsync.cpp
@@ -21,7 +21,7 @@ PStreamEventDispatcher ThreadPoolAsyncImpl::getListener() {
 		cQueue.push(lst);
 		tQueue.push(lst);
 	}
-	while (threadCount.getCounter() < reqThreadCount) {
+	while (threadCount.getCounter() < static_cast<int>(reqThreadCount)) {
 		RefCntPtr<ThreadPoolAsyncImpl> me (this);
 		runThread([me]{me->worker();});
 		threadCount.inc();
@@ -120,9 +120,9 @@ void ThreadPoolAsyncImpl::worker() noexcept {
 
 		t();
 
-		if (threadCount.getCounter() > reqThreadCount) {
+		if (threadCount.getCounter() > static_cast<int>(reqThreadCount)) {
 			Sync _(lock);
-			if (threadCount.getCounter() > reqThreadCount) {
+			if (threadCount.getCounter() > static_cast<int>(reqThreadCount)) {
 				threadCount.dec();
 				return;
 			}
